[rattr] extract RAttrAggregation in separate incluse/source files
