+++ graf2d/gpadv7/inc/ROOT/RAttrAxis.hxx
@@ -1,5 +1,5 @@
 /*************************************************************************
- * Copyright (C) 1995-2019, Rene Brun and Fons Rademakers.               *
+ * Copyright (C) 1995-2021, Rene Brun and Fons Rademakers.               *
  * All rights reserved.                                                  *
  *                                                                       *
  * For the licensing terms see $ROOTSYS/LICENSE.                         *
@@ -9,7 +9,7 @@
 #ifndef ROOT7_RAttrAxis
 #define ROOT7_RAttrAxis
 
-#include <ROOT/RAttrBase.hxx>
+#include <ROOT/RAttrAggregation.hxx>
 #include <ROOT/RAttrLine.hxx>
 #include <ROOT/RAttrText.hxx>
 #include <ROOT/RAttrValue.hxx>
